requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
openai==1.57.4
apscheduler==3.10.4
pydantic==2.10.2
pydantic-settings==2.6.1
dateutils==0.6.12
